bugfix casting (s16) in l3g4is_gyr_get_data function

+++ sysfs/l3gd20h_gyr.c
@@ -711,9 +711,9 @@ static int l3gd20h_gyr_get_data(struct l3gd20h_gyr_status *stat,
 	if (err < 0)
 		return err;
 
-	hw_d[0] = (s32) (((gyro_out[1]) << 8) | gyro_out[0]);
-	hw_d[1] = (s32) (((gyro_out[3]) << 8) | gyro_out[2]);
-	hw_d[2] = (s32) (((gyro_out[5]) << 8) | gyro_out[4]);
+	hw_d[0] = (s32) ((s16)((gyro_out[1]) << 8) | gyro_out[0]);
+	hw_d[1] = (s32) ((s16)((gyro_out[3]) << 8) | gyro_out[2]);
+	hw_d[2] = (s32) ((s16)((gyro_out[5]) << 8) | gyro_out[4]);
 
 	//hw_d[0] = hw_d[0] * stat->sensitivity;
 	//hw_d[1] = hw_d[1] * stat->sensitivity;
